hotfix: changing primary key types
The type for the primary keys in the database is int, but in some entitys the type was bigint, and because of this the code didn't even run

+++ src/domain/entity/energy-bill-group-a.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'conta_energia_grupo_a' })
 export class EnergyBillGroupA {
-  @PrimaryGeneratedColumn({ name: 'conta_energia_a_id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'conta_energia_a_id', type: 'int' })
   id: number;
 
   @Column({ name: 'energia_conta_mes' })
