+++ src/domain/entity/time.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'tempo' })
 export class Time {
-  @PrimaryGeneratedColumn({ name: 'tempo_id', type: 'bigint' })
+  @PrimaryGeneratedColumn({ name: 'tempo_id', type: 'int' })
   id: number;
 
   @Column({ name: 'tempo_mes' })
